requirements.txt
# Core Dependencies
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-socketio==5.3.6
paho-mqtt==1.6.1
kafka-python==2.0.2
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
redis==4.6.0

# Input Validation
pydantic==2.5.0
pydantic[email]==2.5.0
jsonschema==4.20.0

# ML/AI
torch==2.0.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
mlflow==2.7.1
joblib==1.3.2

# Blockchain
web3==6.10.0
eth-account==0.9.0

# Monitoring
prometheus-client==0.17.1
prometheus-flask-exporter==0.23.0

# Security
bcrypt==4.1.2
cryptography==41.0.7
tenacity==8.2.3

# Development
pytest==7.4.2
black==23.7.0
flake8==6.0.0

# New dependencies for enhanced features
graphene==3.3
flask-graphql==2.0.1
psutil==5.9.6
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation-psycopg2==0.41b0
dvc==3.30.0
fastapi==0.104.1
uvicorn==0.24.0

# Simulation
sumolib==1.18.0
traci==1.18.0

# Enhanced ML
scipy==1.11.4
mlflow==2.7.1
sklearn==0.0.post10

# Geospatial
geopy==2.4.0
shapely==2.0.1

# WebSocket support
python-socketio==5.8.0
eventlet==0.33.3

# HTTP Client
requests==2.31.0
httpx==0.25.2